+++ pkg/shelf/example/example_server.dart
@@ -15,5 +15,5 @@ void main() {
 }
 
 shelf.Response _echoRequest(shelf.Request request) {
-  return new shelf.Response.ok('Request for "${request.pathInfo}"');
+  return new shelf.Response.ok('Request for "${request.url}"');
 }